refactor(DDD): déplace shared/ dans infrastructure/
renomme les router.js en routes.js
