+++ src/middlewares/security.js
@@ -2,7 +2,8 @@
 import jwt from 'jsonwebtoken';
 import debugFactory from 'debug';
 
-import APIError from '../shared/APIError.js';
+import APIError from '../infrastructure/shared/APIError.js';
+
 const debug = debugFactory('service:security');
 
 const securityService = {
