+++ src/modules/artworks/controller.js
@@ -4,7 +4,7 @@ import debugFactory from 'debug';
 import dataMapper from '../artworks/model.js';
 import userDataMapper from '../users/model.js';
 import securityService from '../../middlewares/security.js';
-import APIError from '../../shared/APIError.js';
+import APIError from '../../infrastructure/shared/APIError.js';
 const debug = debugFactory('controller');
 
 const artworksController = {
